backend/AI/pipelines/transcriber.py
from groq import Groq
import os
from dotenv import load_dotenv

load_dotenv()
client = Groq(api_key=os.getenv("GROQ_API_KEY"))

def transcribe_audio(audio_file_path: str) -> str:
    """
    Takes path to .mp3 or .wav consultation recording.
    Returns full transcript as string.
    """
    with open(audio_file_path, "rb") as audio_file:
        transcription = client.audio.transcriptions.create(
            model="whisper-large-v3-turbo",
            file=audio_file,
            response_format="text",
            language="en"  # change to "hi" for Hindi, etc.
        )
    return str(transcription)